backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.1
ollama==0.3.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
python-multipart==0.0.9
asyncpg==0.30.0
joblib==1.4.2
google-auth==2.29.0
requests==2.32.3
aiofiles==24.1.0
Pillow==10.4.0
email-validator==2.1.0
